Move plugins data into `fixtures` directory
