+++ client/state/plugins/installed/test/fixtures/plugins.js
